Catch_Leopards-main/backend/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
onnx
opencv-python
numpy
pillow
python-multipart
firebase-admin
python-dotenv
requests
